Fix backup_plan bug
Bug description: When take a break with "Continue After Break" checked, and edit the title of the break, then edit the title of the last plan which is linked to a previous plan and whose plan time is already set to that previous plan, the plan time is remained to that of the previous plan. We expect it to reset to empty.

Reason: the update_link method is called multiple times, and the backup_plan is overridden.

Solution: Don't override backup_plan if the is_continued property is not changed.